Add: use portfolio fetching in different files for easy reading
