+++ src/hooks/usePortfolio/usePortfolioFetch/useCoingeckoFetch.ts
@@ -1,0 +1,73 @@
+// @ts-nocheck TODO: Fill in all missing types before enabling the TS check again
+
+import { useCallback } from 'react'
+
+export default function useCoingeckoFetch({
+    account,
+    currentNetwork,
+    setAssetsByAccount,
+    addToast,
+    getCoingeckoPrices,
+    getCoingeckoPriceByContract,
+    getCoingeckoAssetPlatforms,
+}) {
+
+    const fetchCoingeckoPrices = useCallback(async(tokens, resolve) => {
+        const coingeckoTokensToUpdate = tokens?.filter(token => token.coingeckoId).filter(token => { 
+        if (((new Date().valueOf() - token.priceUpdate ) >= 2*60*1000)) {
+            return token
+        }}).map(token => token.coingeckoId)
+        if (!coingeckoTokensToUpdate.length) return null
+
+        try {
+            const response = await getCoingeckoPrices(coingeckoTokensToUpdate.join(','))
+            if (!response) return null
+            resolve && resolve(response)
+
+        } catch (e) {
+            addToast(e.message, { error: true })
+            resolve && resolve([])
+
+            setAssetsByAccount(prev => ({ ...prev, loading: false }))
+        }
+    }, [account, currentNetwork])
+
+    const fetchCoingeckoAsset = async () => {
+        const network = networks.find(({ id }) => id === currentNetwork)
+        try {
+            const response = await getCoingeckoAssetPlatforms()
+        if (!response) return null
+            const current = response.find(ntw => ntw.chain_identifier === network?.chainId)
+            return current?.id
+        } catch (e) {
+            addToast(e.message, { error: true })
+        }
+    }
+
+    const fetchCoingeckoPricesByContractAddress = useCallback(async(tokens, resolve) => {
+        const assetPlatform = await fetchCoingeckoAsset()
+        const coingeckoTokensToUpdate = tokens.map(token => token.address)
+        try {
+        Promise.all(coingeckoTokensToUpdate.map(async (addr) => {
+            const response = await getCoingeckoPriceByContract(assetPlatform, addr)
+            if (!response) return null
+            return {
+                address: response?.platforms[assetPlatform],
+                tokenImageUrls: response?.image,
+                tokenImageUrl: response?.image?.small,
+                symbol: response?.symbol.toUpperCase(),
+                price: response?.market_data.current_price.usd,
+                isHidden: false,
+            }
+        })).then(res => resolve({ tokens: res, state: 'coingecko' }))
+        } catch (e) {
+            resolve && resolve({ tokens: {}, state: 'coingecko' })
+        }
+    }, [account, currentNetwork])
+
+
+    return {
+        fetchCoingeckoPrices,
+        fetchCoingeckoPricesByContractAddress
+    }
+}
